requirements.txt
streamlit
python-dotenv
transformers
datasets
soundfile
torch
langchain
datasets
soundfile
spotipy
openai
python-dotenv
langchain-community
langchain-anthropic
langchain-google-genai